Changing to decimal
Since postgre doesn't support double, I switch to decimal

+++ db/migrate/20130626221901_add_new_fields_to_celestial_entities.rb
@@ -1,9 +1,9 @@
 class AddNewFieldsToCelestialEntities < ActiveRecord::Migration
   def change
     add_column :celestial_entities, :radius, :integer 
-    add_column :celestial_entities, :volume, :double 
-    add_column :celestial_entities, :density, :double 
-    add_column :celestial_entities, :gravity, :double 
+    add_column :celestial_entities, :volume, :decimal 
+    add_column :celestial_entities, :density, :decimal 
+    add_column :celestial_entities, :gravity, :decimal 
     add_column :celestial_entities, :parent_id, :integer 
     add_column :celestial_entities, :entity_subtypeid, :integer
   end
